Hub: fix knex yarn commands
