--- packages/hub/config/paths.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const path = require('path');
-const fs = require('fs');
-
-// Make sure any symlinks in the project folder are resolved:
-// https://github.com/facebook/create-react-app/issues/637
-const appDirectory = fs.realpathSync(process.cwd());
-const resolveApp = relativePath => path.resolve(appDirectory, relativePath);
-
-// Config after eject: we're in ./config/
-module.exports = {
-  dotenv: resolveApp('.env')
-};
