--- packages/hub/samples/open_channel.ledger.json
@@ -1,30 +1,0 @@
-{
-  "from": "0xffff6147243897776F085aBF5F45F24FC2943669",
-  "commitment": {
-    "allocation": ["0x05", "0x05"],
-    "destination": [
-      "0xffff6147243897776F085aBF5F45F24FC2943669",
-      "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-    ],
-    "channel": {
-      "channelType": "0xabcd10b5ea16F12f5bEFc45d511978CFF2780568",
-      "participants": [
-        "0xffff6147243897776F085aBF5F45F24FC2943669",
-        "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-      ],
-      "nonce": 1000
-    },
-    "turnNum": 0,
-    "appAttributes": "0x00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000000050000000000000000000000000000000000000000000000000000000000000002000000000000000000000000ffff6147243897776f085abf5f45f24fc2943669000000000000000000000000100063c326b27f78b2cbb7cd036b8dde4d4fca7c",
-    "commitmentCount": 0,
-    "commitmentType": 0
-  },
-  "signature": {
-    "message": "0x08c7d9234d31564cfc95bbc57fc84e492e34a91f67ac2b944f702a4417d366e3",
-    "messageHash": "0x15fbb32fa237f2574527aa57b20c2801c1838127fa285f176cc22409148104cd",
-    "v": "0x1b",
-    "r": "0x33fc019b834f43ed07ca60dc47c71ef576c89d80111e153f417a3c617d64d93c",
-    "s": "0x23ea4956cf65b538bf82d6251595a9012c33cec5854f27239a1f3dcb535dd572",
-    "signature": "0x33fc019b834f43ed07ca60dc47c71ef576c89d80111e153f417a3c617d64d93c23ea4956cf65b538bf82d6251595a9012c33cec5854f27239a1f3dcb535dd5721b"
-  }
-}
