--- packages/hub/samples/open_channel.rps.json
@@ -1,29 +1,0 @@
-{
-  "commitment": {
-    "channel": {
-      "channelType": "0xabcd10b5ea16F12f5bEFc45d511978CFF2780568",
-      "participants": [
-        "0xffff6147243897776F085aBF5F45F24FC2943669",
-        "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-      ],
-      "nonce": 1000
-    },
-    "turnNum": 0,
-    "allocation": ["0x05", "0x05"],
-    "destination": [
-      "0xffff6147243897776F085aBF5F45F24FC2943669",
-      "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-    ],
-    "commitmentCount": 0,
-    "commitmentType": 0,
-    "appAttributes": "0x0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-  },
-  "signature": {
-    "message": "0x4a6979dc65c81dec39a8fb9e0a49b216950c359bb91b5bc8fcb70e9f7fcd3476",
-    "messageHash": "0x4cd8f7279aa6efdfd1f3ea11a76a7f0e55230e13ca12bc94b137fff589b18f1b",
-    "v": "0x1c",
-    "r": "0x54d7a906935fea18b48cf0acf7399e75ebb38f2d700314a6c5243eb17ff8ad6a",
-    "s": "0x289a7cc7bd6a806d694c6d2427d2d26e9a3be147c8c5804c0a94c9ba11bd5138",
-    "signature": "0x54d7a906935fea18b48cf0acf7399e75ebb38f2d700314a6c5243eb17ff8ad6a289a7cc7bd6a806d694c6d2427d2d26e9a3be147c8c5804c0a94c9ba11bd51381c"
-  }
-}
