--- packages/hub/samples/update_channel.rps.json
@@ -1,29 +1,0 @@
-{
-  "commitment": {
-    "channel": {
-      "channelType": "0xabcd10b5ea16F12f5bEFc45d511978CFF2780568",
-      "nonce": 1000,
-      "participants": [
-        "0xffff6147243897776F085aBF5F45F24FC2943669",
-        "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-      ]
-    },
-    "turnNum": 2,
-    "allocation": ["0x05", "0x05"],
-    "destination": [
-      "0xffff6147243897776F085aBF5F45F24FC2943669",
-      "0x100063c326b27f78b2cBb7cd036B8ddE4d4FCa7C"
-    ],
-    "commitmentCount": 0,
-    "commitmentType": 1,
-    "appAttributes": "0x0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-  },
-  "signature": {
-    "message": "0x51fcef9afba3febb71cc6daaab8db4b6244d84c24eddbbbaf9681b3ce2fdf924",
-    "messageHash": "0xea9ea3683f252fda7cdcf31975db0a4c36237fa026be4a7e21eab480017701cf",
-    "v": "0x1c",
-    "r": "0x5ec37e7d562fb43ffbb658745a8b6ab6f386bd5bbebd185b375da2c5f1b72f63",
-    "s": "0x21d59df52b05d8abdb347e149b0f2e3d43b2cd1aacfcdfe2403dc1e64b44f8a9",
-    "signature": "0x5ec37e7d562fb43ffbb658745a8b6ab6f386bd5bbebd185b375da2c5f1b72f6321d59df52b05d8abdb347e149b0f2e3d43b2cd1aacfcdfe2403dc1e64b44f8a91c"
-  }
-}
