+++ packages/hub/src/test/test-data.ts
@@ -1,6 +1,5 @@
 import {Channel, encodeConsensusData, getChannelId, State} from '@statechannels/nitro-protocol';
 import {ConsensusData} from '@statechannels/nitro-protocol/lib/src/contract/consensus-data';
-import {HUB_ADDRESS} from '../constants';
 import {
   allocationOutcome2,
   allocationOutcome3,
@@ -12,7 +11,6 @@ import {
   FUNDED_NONCE_3,
   NONCE,
   ONGOING_APP_NONCE,
-  PARTICIPANT_1_ADDRESS,
   PARTICIPANTS,
   PARTICIPANTS_3,
   UNFUNDED_NONCE
@@ -152,64 +150,6 @@ export const stateConstructors = {
   postfundSetup3
 };
 
-const baseResponse = {
-  channel: {
-    channelNonce: expect.any(String),
-    participants: PARTICIPANTS,
-    chainId: DUMMY_CHAIN_ID
-  },
-  outcome: allocationOutcome2,
-  challengeDuration: 1000,
-  appDefinition: DUMMY_RULES_ADDRESS,
-  isFinal: false
-};
-
-const baseResponse3 = {
-  channel: {
-    channelNonce: expect.any(String),
-    participants: PARTICIPANTS_3,
-    chainId: DUMMY_CHAIN_ID
-  },
-  outcome: allocationOutcome3,
-  challengeDuration: 1000,
-  appDefinition: DUMMY_RULES_ADDRESS,
-  isFinal: false
-};
-
-// Ledger Channel Manager Responses
-export const prefundSetup1Response: State = {
-  ...baseResponse,
-  turnNum: 1,
-  appData: encodeConsensusData(consensusAppData2(0))
-};
-
-export const prefundSetup2Response3: State = {
-  ...baseResponse3,
-  turnNum: 2,
-  appData: encodeConsensusData(consensusAppData3(0))
-};
-
-export const postfundSetup1Response: State = {
-  ...baseResponse,
-  turnNum: 3,
-  appData: encodeConsensusData(consensusAppData2(0)),
-  channel: fundedChannel
-};
-
-export const postfundSetup2Response3: State = {
-  ...baseResponse3,
-  turnNum: 5,
-  appData: encodeConsensusData(consensusAppData3(0)),
-  channel: fundedChannel3
-};
-
-export const app1Response: State = {
-  ...baseResponse,
-  turnNum: 5,
-  appData: encodeConsensusData({proposedOutcome: [], furtherVotesRequired: 0}),
-  channel: beginningAppPhaseChannel
-};
-
 // Ledger Channel Manager input states
 export const createdPrefundSetup1: State = {
   channel: defaultChannel,
@@ -229,10 +169,4 @@ export const createdPrefundSetup2Participants3: State = {
   isFinal: false,
   challengeDuration: 1000,
   appDefinition: DUMMY_RULES_ADDRESS
-};
-
-export const createdChannel = {
-  id: expect.any(Number),
-  participants: [{address: PARTICIPANT_1_ADDRESS}, {address: HUB_ADDRESS}],
-  chainId: DUMMY_CHAIN_ID
 };
