+++ packages/hub/src/wallet/services/__test__/channelManager.test.ts
@@ -2,7 +2,7 @@ import {State} from '@statechannels/nitro-protocol';
 import {signState} from '@statechannels/nitro-protocol/lib/src/signatures';
 import {Signature} from 'ethers/utils';
 import {HUB_PRIVATE_KEY} from '../../../constants';
-import {fundedChannel, stateConstructors as testDataConstructors} from '../../../test/test_data';
+import {fundedChannel, stateConstructors as testDataConstructors} from '../../../test/test-data';
 import * as ChannelManager from '../channelManager';
 
 let prefundSetup0: State;
